+++ src/component/actions/SendEmail/SendEmail.test.ts
@@ -1,8 +1,9 @@
-import { FlowDefinition, SendEmail } from '../../../flowTypes';
-import { composeComponentTestUtils, genSendEmailAction } from '../../../testUtils';
+import { SendEmail } from '../../../flowTypes';
+import { composeComponentTestUtils } from '../../../testUtils';
+import { createSendEmailAction } from '../../../testUtils/assetCreators';
 import SendEmailComp from './SendEmail';
 
-const sendEmailAction = genSendEmailAction();
+const sendEmailAction = createSendEmailAction();
 
 const { setup } = composeComponentTestUtils<SendEmail>(SendEmailComp, sendEmailAction);
