+++ src/testUtils/index.tsx
@@ -6,37 +6,9 @@ import { object } from 'prop-types';
 import * as React from 'react';
 import * as config from '../../__test__/config';
 import { ConfigProviderContext, endpointsPT, flowPT, languagesPT } from '../config';
-import { Operators } from '../config/operatorConfigs';
-import { Types } from '../config/typeConfigs';
-import {
-    AnyAction,
-    CallWebhook,
-    Case,
-    ChangeGroups,
-    ContactProperties,
-    Exit,
-    FlowDefinition,
-    FlowEditorConfig,
-    FlowNode,
-    Group,
-    Methods,
-    Router,
-    SendEmail,
-    SendMsg,
-    SetContactProperty,
-    StartFlow,
-    StartFlowArgs,
-    StartFlowExitNames,
-    SwitchRouter,
-    Wait,
-    WaitTypes,
-    Field
-} from '../flowTypes';
-import { AppState, createStore, initialState, SearchResult } from '../store';
-import { capitalize, getBaseLanguage, merge, set } from '../utils';
-import { SetContactField, SetRunResult } from '../flowTypes';
-
-const { results: groupsResults } = require('../../assets/groups.json');
+import { FlowDefinition, FlowEditorConfig } from '../flowTypes';
+import { AppState, createStore, initialState } from '../store';
+import { getBaseLanguage, merge, set } from '../utils';
 
 export interface Resp {
     results: Array<{ [key: string]: any }>;
@@ -176,293 +148,3 @@ export const composeComponentTestUtils = <P extends {}>(
     setup: composeSetup<P>(Component, baseProps, baseDuxState, baseContext),
     spyOn: composeSpy(Component)
 });
-
-/**
- * Create a select control option
- */
-export const genSelectOption = ({ label }: { label: string }) => ({
-    label: capitalize(label.trim()),
-    labelKey: 'name',
-    valueKey: 'id'
-});
-
-export const genSendMsgAction = ({
-    uuid = 'send_msg-0',
-    text = 'Hey!',
-    all_urns = false
-}: {
-    uuid?: string;
-    text?: string;
-    // tslint:disable-next-line:variable-name
-    all_urns?: boolean;
-} = {}): SendMsg => ({
-    type: Types.send_msg,
-    uuid,
-    text,
-    all_urns
-});
-
-export const genSendEmailAction = ({
-    uuid = 'send_email-0',
-    subject = 'New Sign Up',
-    body = '@run.results.name just signed up.',
-    emails = ['jane@exaumple.com']
-}: {
-    uuid?: string;
-    subject?: string;
-    body?: string;
-    emails?: string[];
-} = {}): SendEmail => ({
-    uuid,
-    type: Types.send_email,
-    subject,
-    body,
-    emails
-});
-
-export const genCallWebhookAction = ({
-    uuid = 'call_webhook-0',
-    url = 'https://www.example.com',
-    method = Methods.GET
-}: {
-    uuid?: string;
-    url?: string;
-    method?: Methods;
-} = {}): CallWebhook => ({
-    uuid,
-    type: Types.call_webhook,
-    url,
-    method
-});
-
-export const genAddGroupsAction = ({
-    uuid = 'add_contact_groups-0',
-    groups = groupsResults
-}: { uuid?: string; groups?: Group[] } = {}): ChangeGroups => ({
-    uuid,
-    type: Types.add_contact_groups,
-    groups
-});
-
-export const genStartFlowAction = ({
-    uuid = 'start_flow-0',
-    flow_name = 'Colors',
-    flow_uuid = 'colors-0'
-}: {
-    uuid?: string;
-    // tslint:disable-next-line:variable-name
-    flow_name?: string;
-    // tslint:disable-next-line:variable-name
-    flow_uuid?: string;
-} = {}): StartFlow => ({
-    type: Types.start_flow,
-    uuid: flow_uuid,
-    flow_name: capitalize(flow_name.trim()),
-    flow_uuid
-});
-
-export const genSetContactPropertyAction = ({
-    uuid = 'set_contact_property-0',
-    property = ContactProperties.Email,
-    value = 'jane@example.com'
-}: {
-    uuid?: string;
-    property?: string;
-    value?: string;
-} = {}): SetContactProperty => ({
-    uuid,
-    property,
-    value,
-    type: Types.set_contact_property
-});
-
-export const genSetContactFieldAction = ({
-    uuid = 'set_contact_field-0',
-    field = {
-        key: 'age',
-        name: 'Age'
-    },
-    value = '25'
-}: { uuid?: string; field?: Field; value?: string } = {}): SetContactField => ({
-    uuid,
-    field,
-    value,
-    type: Types.set_contact_field
-});
-
-export const genSetRunResultAction = ({
-    uuid = 'set_run_result-0',
-    result_name = 'Name',
-    value = 'Grace',
-    category = ''
-}: {
-    uuid?: string;
-    result_name?: string;
-    value?: string;
-    category?: string;
-} = {}): SetRunResult => ({
-    uuid,
-    result_name,
-    value,
-    category,
-    type: Types.set_run_result
-});
-
-// tslint:disable-next-line:variable-name
-export const genCase = ({
-    uuid,
-    type,
-    exit_uuid,
-    args = []
-}: {
-    uuid: string;
-    type: Operators;
-    exit_uuid: string;
-    args?: string[];
-}) => ({
-    uuid,
-    type,
-    exit_uuid,
-    arguments: args
-});
-
-export const genExit = ({
-    uuid = 'exit-0',
-    name = null,
-    destination_node_uuid = 'node-1'
-}: {
-    uuid?: string;
-    name?: string;
-    // tslint:disable-next-line:variable-name
-    destination_node_uuid?: string;
-} = {}) => ({
-    uuid,
-    name,
-    destination_node_uuid
-});
-
-// tslint:disable-next-line:variable-name
-export const genWait = (type: WaitTypes, flow_uuid?: string) => ({
-    type,
-    ...(flow_uuid ? { flow_uuid } : {})
-});
-
-// tslint:disable-next-line:variable-name
-export const genRouter = (result_name?: string): Router => ({
-    type: 'switch',
-    ...(result_name ? { result_name } : {})
-});
-
-export const genSwitchRouter = (
-    cases: Case[],
-    operand: string,
-    // tslint:disable-next-line:variable-name
-    default_exit_uuid: string
-) => ({
-    ...genRouter(),
-    cases,
-    operand,
-    default_exit_uuid
-});
-
-export const genFlowNode = ({
-    actions,
-    exits,
-    uuid = 'node-0',
-    router = null,
-    wait = null
-}: {
-    actions: AnyAction[];
-    exits: Exit[];
-    uuid?: string;
-    router?: Router | SwitchRouter;
-    wait?: Wait;
-}): FlowNode => ({
-    actions,
-    exits,
-    uuid,
-    ...(router ? { router } : ({} as any)),
-    ...(wait ? { wait } : ({} as any))
-});
-
-export const genStartFlowNode = (
-    startFlowAction: StartFlow,
-    uuid: string = 'start_flow_node-0',
-    // tslint:disable-next-line:variable-name
-    flow_uuid?: string,
-    exitUUIDs: string[] = ['exit1', 'exit2']
-): FlowNode =>
-    genFlowNode({
-        actions: [startFlowAction],
-        exits: [
-            genExit({
-                uuid: exitUUIDs[0],
-                name: StartFlowExitNames.Complete,
-                destination_node_uuid: 'destination-completed'
-            }),
-            genExit({
-                uuid: exitUUIDs[1],
-                name: StartFlowExitNames.Expired,
-                destination_node_uuid: 'destination-expired'
-            })
-        ],
-        uuid,
-        router: genSwitchRouter(
-            [
-                genCase({
-                    uuid: 'start_flow_case-0',
-                    type: Operators.has_run_status,
-                    exit_uuid: exitUUIDs[0],
-                    args: [StartFlowArgs.Complete]
-                }),
-                genCase({
-                    uuid: 'start_flow_case-1',
-                    type: Operators.has_run_status,
-                    exit_uuid: exitUUIDs[1],
-                    args: [StartFlowArgs.Expired]
-                })
-            ],
-            '@child',
-            null
-        ),
-        wait: genWait(WaitTypes.flow, flow_uuid)
-    });
-
-export const genGroupsRouterNode = (
-    groups: Group[] = groupsResults,
-    uuid: string = 'split_by_groups-0'
-): FlowNode =>
-    genFlowNode({
-        actions: [],
-        exits: groups.map((group, idx) =>
-            genExit({ uuid: group.uuid, name: group.name, destination_node_uuid: `node-${idx}` })
-        ),
-        uuid,
-        router: genSwitchRouter(
-            groups.map((group, idx) =>
-                genCase({
-                    uuid: `split_by_group-${idx}`,
-                    type: Operators.has_group,
-                    exit_uuid: group.uuid,
-                    args: [group.uuid]
-                })
-            ),
-            '@contact.groups',
-            null
-        ),
-        wait: genWait(WaitTypes.group)
-    });
-
-export const getGroupOptions = (groups: Group[] = groupsResults): SearchResult[] =>
-    groups.map(({ name, uuid }) => ({
-        name,
-        id: uuid
-    }));
-
-export const getGroups = (sliceAt: number, groups: Group[] = groupsResults) =>
-    groups
-        .map(({ name, uuid }) => ({
-            name,
-            id: uuid
-        }))
-        .slice(sliceAt);
